backend/requirements.txt
channels==4.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
pydantic==2.8.2
pymongo==4.6.2
python-dotenv==1.0.1
